+++ packages/mel-components/src/mel-select/index.ts
@@ -1,5 +1,9 @@
 import MelSelect from './src/index.vue';
 
+export type { MelSelectProps } from './src/types';
+
+export type MelSelectInstance = InstanceType<typeof MelSelect>;
+
 export {
   MelSelect,
 };