+++ packages/mel-components/src/mel-select/src/index.vue
@@ -4,9 +4,11 @@
 
 <script lang="tsx" setup>
   import type { ISelectProps } from 'element-plus';
+  import type { Ref } from 'vue';
   import type { MelSelectProps, MelSelectSlots, SelectInstance } from './types';
+  import { useOptionsApi } from '@mixte/mel-components/utils';
   import { ElOption, ElSelect } from 'element-plus';
-  import { ref, type Ref, useAttrs } from 'vue';
+  import { computed, ref, toRef, useAttrs } from 'vue';
 
   const props = defineProps<MelSelectProps>();
   const slots = defineSlots<MelSelectSlots>();
@@ -16,16 +18,23 @@
 
   const selectRef = ref<SelectInstance>();
 
+  const { propApi, api, loading } = useOptionsApi(toRef(props, 'optionsApi'));
+
+  const options = computed(() => {
+    return (propApi.value ? api.response : props.options);
+  });
+
   function Render() {
     return (
       <ElSelect
         {...attrs}
         ref={selectRef}
         v-model={value.value}
+        loading={props.loading || loading.value}
       >
         {{
           default: () => {
-            return props.options?.map((option) => {
+            return options.value?.map((option) => {
               return <ElOption key={`${option.value}`} {...option} />;
             });
           },
@@ -35,9 +44,8 @@
     );
   }
 
-  defineExpose<{
-    selectRef: Ref<SelectInstance | undefined>;
-  }>({
-    selectRef,
+  defineExpose({
+    selectRef: selectRef as Ref<SelectInstance | undefined>,
+    api,
   });
 </script>